ASoC: rsnd: Allow config of audio_clk_div3 divider via dt
Add the custom "cetitec,audio-clk-div" property to the rcar-sound device
tree binding. It allows choosing the divisor on the audio_clk_div3[*]
outputs of the audio_clk_div3 block in the ADG module.
These drive the audio_clk_div.[0-2] inputs of the avb-adg-sync multiplexer
with the AUDIO_CLK[ABC] inputs (clk_a, clk_b and clk_c in the dt)divided by
the specified value. (See datasheet section 43.1.2 for an overview).