ASoC: rsnd: Fix handling of SSI0 when in SSI012[9] multi-SSI mode
When the SSI0 is the parent of a multi-SSI setup, it must be fully setup,
started and stopped since it is also part of the playback/capture setup. So
only skip the SSI (as per bc9e2e6c7c40ea8de386560d9f7cf71c45b3b384) if the
SSI is parent outside of a multi-SSI setup (e.g. SSI0 playback, SSI1
capture with shared SCK0/WS0).