ASoC: rsnd: Override SSI clock source selection
Patch the value of the SSI Clock Source Override control added in the prev.
commit through to the rsnd_adg_clk_query function. If the value of the
control is not auto, only the clock source set by the override control will
be considered.
If the desired rate cannot be generated from the requested clock source,
stream setup will fail.