ASoC: rsnd: Add "explicit-only-clocks" to prevent use of HDMI in clock by default
The ADG driver automatically selects the "best" clock for 44100 and 48000Hz
audio playback from the configured input clocks. When one of the input
clocks is a special-purpose clock, e.g. the recovered HDMI input audio
clock from the adv748x in the CETiBox, this could result in nonfunctional
analog audio when the special-purpose clock was selected as the best clock.
Add an explicit-only-clocks attribute which blacklists these clocks from
the automatic configuration in the ADG module. These clocks are then only
used when selected with the "SSI Clock Source Override" control.