ASoC: rsnd: Fix multi-SSI detection in hw rules
The rsnd_runtime_channel_for_ssi_with_params helper used by the hw rules
for channels and rate incorrectly called the rsnd_runtime_is_ssi_multi
helper which expects the params to be installed in the runtime already,
which is not the case when the hw rules are being processed.

Add a rsnd_runtime_is_ssi_multi_with_params helper that, like the other
..._with_params helpers, uses the passed parameters instead.

This fixes the issue that valid rate,channel combinations with more than
2 channels were rejected for multi-SSI devices because the master SSI
was queried for a valid clock with n>2 channels, which may not exist.
With this patch only valid clocks for 2 channels are queried, as each
indivdual SSI in the multi-SSI setup only runs with 2 channels.

Signed-off-by: Matthias Blankertz <matthias.blankertz@cetitec.com>