drm: dw-hdmi-i2s: Workaround for interchannel sample offset
Add a second reset of the I2S block after the audio parameters are set.
Without this, it was observed on the R-Car Gen3 platform that in ~50% of
audio stream starts, the samples on each odd-numbered channel were
delayed by one sample relative to the even-numbered channels.

Maybe this relates to the comment of commit f3d0af1a4cea ("drm/bridge:
dw-hdmi-i2s: reset audio fifo before applying new params"), but without
the datasheet it is only speculation.

Signed-off-by: Matthias Blankertz <matthias.blankertz@cetitec.com>