ASoC: rsnd: Restrict use of shared SSI
Restrict the reuse of a opened SSI to child <-> parent relationships
and multiple DAIs attached to the same SSI via MIX block. In other
cases, block the reuse with EBUSY.

This allows using the same SSI for multiple different DAI setups, e.g.
a DAI for analog audio I/O using SSI0 and SSI1, and a HDMI audio output
using SSI0,1,2,9. Userspace can then choose which output to use, and the
other DAI will then return EBUSY until the audio stream is closed.

Before this fix, this setup resulted in crackling, distortion in the
running stream, or the SSI could hang, when the second DAI was started and
interfered with the running SSI.

Signed-off-by: Matthias Blankertz <matthias.blankertz@cetitec.com>