ravb: Disable LINK interrupt in hw when renesas,no-ether-link is set
The renesas,no-ether-link device tree setting disables the AVB_LINK
input and link state changes on LINK interrupts. This is used for
example when the AVB_LINK pin is not used in this function on the board.

The driver unconditionally enabled the interrupt and considered the
no-ether-link setting (there called no_avb_link) in the interrupt
handler:
   static void ravb_emac_interrupt_unlocked(...) {
      [...]
      if (ecsr & ECSR_LCHNG) {
         /* Link changed */
         if (priv->no_avb_link)
            return;
      [...]
On a Renesas "Vehicle Computer 2" board, where this pin is unconnected,
it was observed that excessive (millions) EMAC interrupts occured, which
affected system performance.

This fixes the driver to disable the LINK interrupt of the EMAC (LCHNGIP
bit, called LINKIM in the datasheet) in hardware when the device tree is
configured with renesas,no-ether-link.

Signed-off-by: Matthias Blankertz <matthias.blankertz@cetitec.com>